move log.py to geneditid, rename python/scripts to python/geneditidtools, rename shell/ to scripts/ and rename job_* scripts to run_*
