+++ TODO.md
@@ -122,7 +122,7 @@ source ~/GenEditID/venv/bin/activate
   chr16_53704130,TCCAGGGCGAGGGATCTAC,GAAGGGCTTGGTTTGATGGC,TCCAGGGCGAGGGATCTACGCAGCTTGCGGTGGCGAAGGCGGCTTTAGTGGCAGCATGAAGCGCACCCCGACTGCCGAGGAACGAGAGCGCGAAGCTAAGGTATGTCGGGCTCCCGGGGCCTGGAGATCTTCGTGCGCTGTGAGCAAGGATCAGGGAACCGGAAGGGCTTGGTTTGATGGC,chr16:53704130-53704310
   chr16_54931181,CCATGCCCGTGTGTG,GCTACAACTCGCACCTCCA,CCATGCCCGTGTGTGGCCATGTCCTATCCGCAGGGCTACTTGTACCAGCCGTCCGCCTCGCTGGCGCTCTACTCGTGCCCGGCGTACAGCACCAGCGTCATTTCGGGGCCCCGCACGGATGAGCTCGGCCGCTCTTCTTCGGGCTCCGCGTTCTCGCCCTACGCTGGCTCGACTGCCTTCACGGCGCCCTCGCCGGGCTACAACTCGCACCTCCA,chr16:54931181-54931395
   ```
-- Run [geneditid_run_amplicount](https://github.com/GenEditID/GenEditID/blob/master/python/scripts/run_ampli_count.py) script on all fasta files from your project directory
+- Run [geneditid_run_amplicount](https://github.com/GenEditID/GenEditID/blob/master/python/geneditidtools/run_ampli_count.py) script on all fasta files from your project directory
   ```
   geneditid_run_amplicount --fastqdir=fastq/
   ```
@@ -149,7 +149,7 @@ source ~/GenEditID/venv/bin/activate
 
 ## Step 5: Identify variants and plot results
 
-- Run [geneditid_run_variantid](https://github.com/GenEditID/GenEditID/blob/master/python/scripts/run_variant_id.py) script from your project directory:
+- Run [geneditid_run_variantid](https://github.com/GenEditID/GenEditID/blob/master/python/geneditidtools/run_variant_id.py) script from your project directory:
 ```
 geneditid_run_variantid
 INFO:pyensembl.sequence_data:Loaded sequence dictionary from /Users/pajanne/Library/Caches/pyensembl/GRCh38/ensembl95/Homo_sapiens.GRCh38.cdna.all.fa.gz.pickle
