+++ data/GEP00001/GEP00001.md
@@ -14,8 +14,8 @@ scp shell/ngs/* clust1-headnode:/scratchb/bioinformatics/pajon01/genome-editing/
 
 scp clust1-headnode:/scratchb/bioinformatics/pajon01/genome-editing/GEP00001/samplesheet.csv .
 scp clust1-headnode:/scratchb/bioinformatics/pajon01/genome-editing/GEP00001/filelist.csv .
-python python/scripts/convert_filelist.py > ~/mnt/scratchb/genome-editing/GEP00001/filelist.csv
-python python/scripts/convert_samplesheet.py  > ~/mnt/scratchb/genome-editing/GEP00001/samplesheet.csv
+python python/geneditidtools/convert_filelist.py > ~/mnt/scratchb/genome-editing/GEP00001/filelist.csv
+python python/geneditidtools/convert_samplesheet.py  > ~/mnt/scratchb/genome-editing/GEP00001/samplesheet.csv
 ```
 
 ```
@@ -34,7 +34,7 @@ tail -f joinreads.out
 ## Read counts
 
 ```
-python python/scripts/create_pipeline_files.py --project=GEP00001 --genome=/Users/pajon01/mnt/refdata/reference_genomes/homo_sapiens/GRCh38_hs38d1/fasta/hsa.GRCh38_hs38d1.fa --seq-dict=/Users/pajon01/mnt/refdata/reference_genomes/homo_sapiens/GRCh38_hs38d1/fasta/hsa.GRCh38_hs38d1.dict --filelist=/Users/pajon01/mnt/scratchb/genome-editing/GEP00001/filelist.csv
+python python/geneditidtools/create_pipeline_files.py --project=GEP00001 --genome=/Users/pajon01/mnt/refdata/reference_genomes/homo_sapiens/GRCh38_hs38d1/fasta/hsa.GRCh38_hs38d1.fa --seq-dict=/Users/pajon01/mnt/refdata/reference_genomes/homo_sapiens/GRCh38_hs38d1/fasta/hsa.GRCh38_hs38d1.dict --filelist=/Users/pajon01/mnt/scratchb/genome-editing/GEP00001/filelist.csv
 
 scp targets.txt clust1-headnode:/scratchb/bioinformatics/pajon01/genome-editing/GEP00001/.
 scp amplicons.txt clust1-headnode:/scratchb/bioinformatics/pajon01/genome-editing/GEP00001/.
@@ -48,7 +48,7 @@ sbatch job_amplicount.sh
 tail -f amplicount.out
 
 source /home/pajon01/genome-editing/venv/bin/activate
-python /home/pajon01/genome-editing/python/scripts/ampliplot.py
+python /home/pajon01/genome-editing/python/geneditidtools/ampliplot.py
 ```
 
 ## MultiQC report
@@ -76,7 +76,7 @@ tail -f amplicon-vardict.out
 
 Config file generator:
 ```
-python python/scripts/create_amplican_conf.py --project=GEP00001 --genome=/Users/pajon01/mnt/refdata/reference_genomes/homo_sapiens/GRCh38_hs38d1/fasta/hsa.GRCh38_hs38d1.fa
+python python/geneditidtools/create_amplican_conf.py --project=GEP00001 --genome=/Users/pajon01/mnt/refdata/reference_genomes/homo_sapiens/GRCh38_hs38d1/fasta/hsa.GRCh38_hs38d1.fa
 ```
 
 ## Read counts - consequence categorisation
@@ -86,5 +86,5 @@ scp clust1-headnode.cri.camres.org:/scratchb/bioinformatics/pajon01/genome-editi
 scp clust1-headnode.cri.camres.org:/scratchb/bioinformatics/pajon01/genome-editing/GEP00001/amplicount.csv .
 scp clust1-headnode.cri.camres.org:/scratchb/bioinformatics/pajon01/genome-editing/GEP00001/amplicount_config.csv .
 
-python python/scripts/ampliplot.py
+python python/geneditidtools/ampliplot.py
 ```
