+++ data/GEP00005/GEP00005.md
@@ -95,7 +95,7 @@ chr2	44344474	44344605	+	GRCh38_chr2_44344474
 ```
 
 ```
-python python/scripts/amplifind.py GEP00005
+python python/geneditidtools/amplifind.py GEP00005
 ```
 
 - New reference genome, need to be re-align
@@ -115,10 +115,10 @@ scp shell/ngs/* clust1-headnode:/scratchb/bioinformatics/pajon01/genome-editing/
 
 ```
 cp ~/mnt/scratchb/genome-editing/GEP00005/samplesheet.csv .
-python python/scripts/convert_samplesheet.py > ~/mnt/scratchb/genome-editing/GEP00005/samplesheet.csv
+python python/geneditidtools/convert_samplesheet.py > ~/mnt/scratchb/genome-editing/GEP00005/samplesheet.csv
 
 cp ~/mnt/scratchb/genome-editing/GEP00005/filelist.csv .
-python python/scripts/convert_filelist.py > ~/mnt/scratchb/genome-editing/GEP00005/filelist.csv
+python python/geneditidtools/convert_filelist.py > ~/mnt/scratchb/genome-editing/GEP00005/filelist.csv
 ```
 
 ```
@@ -128,7 +128,7 @@ sbatch job_alignment.sh
 - Re-run pipeline
 
 ```
-python python/scripts/create_pipeline_files.py --project=GEP00005 --genome=/Users/pajon01/mnt/refdata/reference_genomes/homo_sapiens/GRCh38_hs38d1/fasta/hsa.GRCh38_hs38d1.fa --seq-dict=/Users/pajon01/mnt/refdata/reference_genomes/homo_sapiens/GRCh38_hs38d1/fasta/hsa.GRCh38_hs38d1.dict --filelist=/Users/pajon01/mnt/scratchb/genome-editing/GEP00005/filelist.csv
+python python/geneditidtools/create_pipeline_files.py --project=GEP00005 --genome=/Users/pajon01/mnt/refdata/reference_genomes/homo_sapiens/GRCh38_hs38d1/fasta/hsa.GRCh38_hs38d1.fa --seq-dict=/Users/pajon01/mnt/refdata/reference_genomes/homo_sapiens/GRCh38_hs38d1/fasta/hsa.GRCh38_hs38d1.dict --filelist=/Users/pajon01/mnt/scratchb/genome-editing/GEP00005/filelist.csv
 scp targets.txt clust1-headnode:/scratchb/bioinformatics/pajon01/genome-editing/GEP00005/.
 scp amplicons.txt clust1-headnode:/scratchb/bioinformatics/pajon01/genome-editing/GEP00005/.
 scp samples.txt clust1-headnode:/scratchb/bioinformatics/pajon01/genome-editing/GEP00005/.
@@ -143,7 +143,7 @@ sbatch job_amplicon_vardict.sh
 ## Read counts
 
 ```
-python python/scripts/create_pipeline_files.py --project=GEP00005 --genome=/Users/pajon01/mnt/refdata/reference_genomes/homo_sapiens/GRCh38_hs38d1/fasta/hsa.GRCh38_hs38d1.fa --seq-dict=/Users/pajon01/mnt/refdata/reference_genomes/homo_sapiens/GRCh38_hs38d1/fasta/hsa.GRCh38_hs38d1.dict --filelist=/Users/pajon01/mnt/scratchb/genome-editing/GEP00005/filelist.csv
+python python/geneditidtools/create_pipeline_files.py --project=GEP00005 --genome=/Users/pajon01/mnt/refdata/reference_genomes/homo_sapiens/GRCh38_hs38d1/fasta/hsa.GRCh38_hs38d1.fa --seq-dict=/Users/pajon01/mnt/refdata/reference_genomes/homo_sapiens/GRCh38_hs38d1/fasta/hsa.GRCh38_hs38d1.dict --filelist=/Users/pajon01/mnt/scratchb/genome-editing/GEP00005/filelist.csv
 
 scp amplicount_config.csv clust1-headnode:/scratchb/bioinformatics/pajon01/genome-editing/GEP00005/.
 scp shell/ngs/job_amplicount.sh clust1-headnode:/scratchb/bioinformatics/pajon01/genome-editing/GEP00005/.
@@ -154,7 +154,7 @@ sbatch job_amplicount.sh
 tail -f amplicount.out
 
 source /home/pajon01/genome-editing/venv/bin/activate
-python /home/pajon01/genome-editing/python/scripts/ampliplot.py
+python /home/pajon01/genome-editing/python/geneditidtools/ampliplot.py
 ```
 
 ## Read counts - consequence categorisation
@@ -164,5 +164,5 @@ scp clust1-headnode.cri.camres.org:/scratchb/bioinformatics/pajon01/genome-editi
 scp clust1-headnode.cri.camres.org:/scratchb/bioinformatics/pajon01/genome-editing/GEP00005/amplicount.csv .
 scp clust1-headnode.cri.camres.org:/scratchb/bioinformatics/pajon01/genome-editing/GEP00005/amplicount_config.csv .
 
-python python/scripts/ampliplot.py
+python python/geneditidtools/ampliplot.py
 ```
