+++ data/GEP00009/GEP00009.md
@@ -50,7 +50,7 @@ chr16	54931181	54931395	+	GRCh38_chr16_54931181
 ```
 
 ```
-python python/scripts/amplifind.py GEP00009
+python python/geneditidtools/amplifind.py GEP00009
 ```
 
 - New reference genome, need to be re-align
@@ -66,10 +66,10 @@ scp shell/ngs/* clust1-headnode:/scratchb/bioinformatics/pajon01/genome-editing/
 
 ```
 cp ~/mnt/scratchb/genome-editing/GEP00009/samplesheet.csv .
-python python/scripts/convert_samplesheet.py > ~/mnt/scratchb/genome-editing/GEP00009/samplesheet.csv
+python python/geneditidtools/convert_samplesheet.py > ~/mnt/scratchb/genome-editing/GEP00009/samplesheet.csv
 
 cp ~/mnt/scratchb/genome-editing/GEP00009/filelist.csv .
-python python/scripts/convert_filelist.py > ~/mnt/scratchb/genome-editing/GEP00009/filelist.csv
+python python/geneditidtools/convert_filelist.py > ~/mnt/scratchb/genome-editing/GEP00009/filelist.csv
 ```
 
 ```
@@ -79,7 +79,7 @@ sbatch job_alignment.sh
 - Re-run pipeline
 
 ```
-python python/scripts/create_pipeline_files.py --project=GEP00009 --genome=/Users/pajon01/mnt/refdata/reference_genomes/homo_sapiens/GRCh38_hs38d1/fasta/hsa.GRCh38_hs38d1.fa --seq-dict=/Users/pajon01/mnt/refdata/reference_genomes/homo_sapiens/GRCh38_hs38d1/fasta/hsa.GRCh38_hs38d1.dict --filelist=/Users/pajon01/mnt/scratchb/genome-editing/GEP00009/filelist.csv
+python python/geneditidtools/create_pipeline_files.py --project=GEP00009 --genome=/Users/pajon01/mnt/refdata/reference_genomes/homo_sapiens/GRCh38_hs38d1/fasta/hsa.GRCh38_hs38d1.fa --seq-dict=/Users/pajon01/mnt/refdata/reference_genomes/homo_sapiens/GRCh38_hs38d1/fasta/hsa.GRCh38_hs38d1.dict --filelist=/Users/pajon01/mnt/scratchb/genome-editing/GEP00009/filelist.csv
 scp targets.txt clust1-headnode:/scratchb/bioinformatics/pajon01/genome-editing/GEP00009/.
 scp amplicons.txt clust1-headnode:/scratchb/bioinformatics/pajon01/genome-editing/GEP00009/.
 scp samples.txt clust1-headnode:/scratchb/bioinformatics/pajon01/genome-editing/GEP00009/.
@@ -103,7 +103,7 @@ sbatch job_amplicon_gatk.sh
 ## Read counts
 
 ```
-python python/scripts/create_pipeline_files.py --project=GEP00009 --genome=/Users/pajon01/mnt/refdata/reference_genomes/homo_sapiens/GRCh38_hs38d1/fasta/hsa.GRCh38_hs38d1.fa --seq-dict=/Users/pajon01/mnt/refdata/reference_genomes/homo_sapiens/GRCh38_hs38d1/fasta/hsa.GRCh38_hs38d1.dict --filelist=/Users/pajon01/mnt/scratchb/genome-editing/GEP00009/filelist.csv
+python python/geneditidtools/create_pipeline_files.py --project=GEP00009 --genome=/Users/pajon01/mnt/refdata/reference_genomes/homo_sapiens/GRCh38_hs38d1/fasta/hsa.GRCh38_hs38d1.fa --seq-dict=/Users/pajon01/mnt/refdata/reference_genomes/homo_sapiens/GRCh38_hs38d1/fasta/hsa.GRCh38_hs38d1.dict --filelist=/Users/pajon01/mnt/scratchb/genome-editing/GEP00009/filelist.csv
 
 scp amplicount_config.csv clust1-headnode:/scratchb/bioinformatics/pajon01/genome-editing/GEP00009/.
 scp shell/ngs/job_amplicount.sh clust1-headnode:/scratchb/bioinformatics/pajon01/genome-editing/GEP00009/.
@@ -114,7 +114,7 @@ sbatch job_amplicount.sh
 tail -f amplicount.out
 
 source /home/pajon01/genome-editing/venv/bin/activate
-python /home/pajon01/genome-editing/python/scripts/ampliplot.py
+python /home/pajon01/genome-editing/python/geneditidtools/ampliplot.py
 
 ```
 
@@ -125,5 +125,5 @@ scp clust1-headnode.cri.camres.org:/scratchb/bioinformatics/pajon01/genome-editi
 scp clust1-headnode.cri.camres.org:/scratchb/bioinformatics/pajon01/genome-editing/GEP00009/amplicount.csv .
 scp clust1-headnode.cri.camres.org:/scratchb/bioinformatics/pajon01/genome-editing/GEP00009/amplicount_config.csv .
 
-python python/scripts/ampliplot.py
+python python/geneditidtools/ampliplot.py
 ```
