+++ data/GEP00010/GEP00010.md
@@ -50,7 +50,7 @@ chr4	26481537	26481784	-	GRCh38_chr4_26481537
 ```
 
 ```
-python python/scripts/amplifind.py GEP00010
+python python/geneditidtools/amplifind.py GEP00010
 ```
 
 - Re-run pipeline
@@ -78,10 +78,10 @@ head -1 filelist.csv
 
 ```
 cp ~/mnt/scratchb/genome-editing/GEP00010/samplesheet.csv .
-python python/scripts/convert_samplesheet.py > ~/mnt/scratchb/genome-editing/GEP00010/samplesheet.csv
+python python/geneditidtools/convert_samplesheet.py > ~/mnt/scratchb/genome-editing/GEP00010/samplesheet.csv
 
 cp ~/mnt/scratchb/genome-editing/GEP00010/filelist.csv .
-python python/scripts/convert_filelist.py > ~/mnt/scratchb/genome-editing/GEP00010/filelist.csv
+python python/geneditidtools/convert_filelist.py > ~/mnt/scratchb/genome-editing/GEP00010/filelist.csv
 ```
 
 ```
@@ -89,7 +89,7 @@ sbatch job_alignment.sh
 ```
 
 ```
-python python/scripts/create_pipeline_files.py --project=GEP00010 --genome=/Users/pajon01/mnt/refdata/reference_genomes/homo_sapiens/GRCh38_hs38d1/fasta/hsa.GRCh38_hs38d1.fa --seq-dict=/Users/pajon01/mnt/refdata/reference_genomes/homo_sapiens/GRCh38_hs38d1/fasta/hsa.GRCh38_hs38d1.dict --filelist=/Users/pajon01/mnt/scratchb/genome-editing/GEP00010/filelist.csv
+python python/geneditidtools/create_pipeline_files.py --project=GEP00010 --genome=/Users/pajon01/mnt/refdata/reference_genomes/homo_sapiens/GRCh38_hs38d1/fasta/hsa.GRCh38_hs38d1.fa --seq-dict=/Users/pajon01/mnt/refdata/reference_genomes/homo_sapiens/GRCh38_hs38d1/fasta/hsa.GRCh38_hs38d1.dict --filelist=/Users/pajon01/mnt/scratchb/genome-editing/GEP00010/filelist.csv
 scp targets.txt clust1-headnode:/scratchb/bioinformatics/pajon01/genome-editing/GEP00010/.
 scp amplicons.txt clust1-headnode:/scratchb/bioinformatics/pajon01/genome-editing/GEP00010/.
 scp samples.txt clust1-headnode:/scratchb/bioinformatics/pajon01/genome-editing/GEP00010/.
@@ -110,9 +110,9 @@ sshfs pajon01@clust1-headnode.cri.camres.org:/mnt/scratchb/bioinformatics/pajon0
 
 ll /Users/pajon01/mnt/scratchb/genome-editing/GEP00010
 cd /Users/pajon01/workspace/genome-editing
-python python/scripts/amplicount.py --fastqdir=/Users/pajon01/mnt/scratchb/genome-editing/GEP00010/fastq/
+python python/geneditidtools/amplicount.py --fastqdir=/Users/pajon01/mnt/scratchb/genome-editing/GEP00010/fastq/
 
-# config file generated with python/scripts/create_pipeline_files.py
+# config file generated with python/geneditidtools/create_pipeline_files.py
 scp amplicount_config.csv clust1-headnode:/scratchb/bioinformatics/pajon01/genome-editing/GEP00010/.
 scp shell/ngs/job_amplicount.sh clust1-headnode:/scratchb/bioinformatics/pajon01/genome-editing/GEP00010/.
 
@@ -123,7 +123,7 @@ sbatch job_amplicount.sh
 tail -f amplicount.out
 
 source /home/pajon01/genome-editing/venv/bin/activate
-python /home/pajon01/genome-editing/python/scripts/ampliplot.py
+python /home/pajon01/genome-editing/python/geneditidtools/ampliplot.py
 
 # umount
 umount /Users/pajon01/mnt/scratchb
@@ -144,7 +144,7 @@ scp clust1-headnode.cri.camres.org:/scratchb/bioinformatics/pajon01/genome-editi
 scp clust1-headnode.cri.camres.org:/scratchb/bioinformatics/pajon01/genome-editing/GEP00010/amplicount.csv .
 scp clust1-headnode.cri.camres.org:/scratchb/bioinformatics/pajon01/genome-editing/GEP00010/amplicount_config.csv .
 
-python python/scripts/ampliplot.py
+python python/geneditidtools/ampliplot.py
 ```
 
 ## Read counts - rare sequence analysis
