+++ python/geneditidtools/delete_project.py
@@ -1,29 +1,28 @@
 import os
 import argparse
-import log as logger
 
+import geneditid.log as logger
 from geneditid.config import cfg
 from geneditid.connect import dbsession
-
-from geneditid.plotter import Plotter
-
+from geneditid.loader import ProjectLoader
 
 def main():
     parser = argparse.ArgumentParser()
     parser.add_argument("--geid", dest="geid", action="store", help="Project GEP ID e.g. 'GEP00001'", required=True)
     options = parser.parse_args()
 
-    log = logger.get_custom_logger(os.path.join(cfg['PROJECTS_FOLDER'], 'run_variandid.log'))
+    log = logger.get_custom_logger(os.path.join(cfg['PROJECTS_FOLDER'], 'delete_project.log'))
 
     try:
-        plotter = Plotter(dbsession, options.geid)
-        plotter.coverage_plot()
-        plotter.variant_impact_plot()
-        #plotter.heatmap_plot()
+        project = ProjectLoader(dbsession)
+        project.delete_project(options.geid)
+        dbsession.commit()
     except Exception as e:
-        logging.exception(e)
+        log.exception(e)
+        dbsession.rollback()
     finally:
         dbsession.close()
+
 
 if __name__ == '__main__':
     main()
