+++ python/geneditidtools/load_ref_data.py
@@ -1,6 +1,6 @@
 import os
-import log as logger
 
+import geneditid.log as logger
 from geneditid.config import cfg
 from geneditid.connect import dbsession
 from geneditid.loader import RefLoader
