+++ python/geneditidtools/run_ampli_count.py
@@ -2,11 +2,11 @@ import os
 import gzip
 import argparse
 import collections
-import log as logger
 
 import pandas as pd
 from Bio import SeqIO
 
+import geneditid.log as logger
 from geneditid.config import cfg
