+++ python/geneditidtools/run_ampli_find.py
@@ -1,25 +1,25 @@
 import os
 import argparse
-import log as logger
 
+import geneditid.log as logger
 from geneditid.config import cfg
 from geneditid.connect import dbsession
-from geneditid.loader import ProjectLoader
+from geneditid.finder import AmpliconFinder
+
 
 def main():
     parser = argparse.ArgumentParser()
     parser.add_argument("--geid", dest="geid", action="store", help="Project GEP ID e.g. 'GEP00001'", required=True)
     options = parser.parse_args()
 
-    log = logger.get_custom_logger(os.path.join(cfg['PROJECTS_FOLDER'], 'delete_project.log'))
+    log = logger.get_custom_logger(os.path.join(cfg['PROJECTS_FOLDER'], 'run_ampli_find.log'))
 
     try:
-        project = ProjectLoader(dbsession)
-        project.delete_project(options.geid)
-        dbsession.commit()
+        log.info('Searching amplicon sequences for project ID\t{}'.format(options.geid))
+        finder = AmpliconFinder(dbsession, options.geid)
+        finder.write_amplicount_config_file()
     except Exception as e:
-        log.exception(e)
-        dbsession.rollback()
+        logging.exception(e)
     finally:
         dbsession.close()
