+++ python/geneditidtools/run_variant_id.py
@@ -1,10 +1,11 @@
 import os
 import argparse
-import log as logger
 
+import geneditid.log as logger
 from geneditid.config import cfg
 from geneditid.connect import dbsession
-from geneditid.finder import AmpliconFinder
+
+from geneditid.plotter import Plotter
 
 
 def main():
@@ -12,17 +13,17 @@ def main():
     parser.add_argument("--geid", dest="geid", action="store", help="Project GEP ID e.g. 'GEP00001'", required=True)
     options = parser.parse_args()
 
-    log = logger.get_custom_logger(os.path.join(cfg['PROJECTS_FOLDER'], 'run_ampli_find.log'))
+    log = logger.get_custom_logger(os.path.join(cfg['PROJECTS_FOLDER'], 'run_variandid.log'))
 
     try:
-        log.info('Searching amplicon sequences for project ID\t{}'.format(options.geid))
-        finder = AmpliconFinder(dbsession, options.geid)
-        finder.write_amplicount_config_file()
+        plotter = Plotter(dbsession, options.geid)
+        plotter.coverage_plot()
+        plotter.variant_impact_plot()
+        #plotter.heatmap_plot()
     except Exception as e:
         logging.exception(e)
     finally:
         dbsession.close()
-
 
 if __name__ == '__main__':
     main()
