+++ scripts/run_fastqc.sh
