+++ scripts/run_joinreads.sh
