+++ scripts/run_mergereads.sh
